backend/pyproject.toml
[tool.poetry]
name = "plant_api"
version = "0.1.0"
description = "FastAPI backend for the plant app"
authors = ["dantheand <danderson@bighatbio.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.25.0"
mangum = "^0.17.0"
python-multipart = "^0.0.6"
pydantic = "^2.5.3"
authlib = "^1.3.0"
google-auth-oauthlib = "^1.2.0"
pillow = "^10.2.0"
fastapi = {extras = ["all"], version = "^0.108.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
urllib3 = ">=1.25.4,<1.27"
aioboto3 = "^12.2.0"
pytest-mock = "^3.12.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^7.4.4"
moto = "^4.2.13"
faker = "^22.1.0"
mypy = "^1.8.0"
types-pillow = "^10.1.0.20240106"
types-python-jose = "^3.3.4.20240106"
boto3-stubs = "^1.34.15"
google-auth-stubs = "^0.2.0"
types-ujson = "^5.9.0.0"
pytest-asyncio = "^0.23.4"
types-aioboto3 = "^12.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
